Refactor SDE tests and update dependencies in Telescope
- Removed the integration test file for SdeManager.
- Updated dependencies in Cargo.toml for Telescope, including eframe and egui_extras.
- Changed factor type from i64 to f64 in various structs and methods to improve precision.
- Adjusted calculations in UniversePane and RegionPane to accommodate the new factor type.
- Fixed issues with region coordinates and system connections in the Telescope app.
-Started the migration of sde from local codebase to library crate

+++ crates/telescope/src/app.rs
@@ -1146,7 +1146,7 @@ impl TelescopeApp {
     fn generate_pane(
         receiver: BCReceiver<MapSync>,
         path: PathBuf,
-        factor: i64,
+        factor: f64,
         region_id: Option<usize>,
         task_msg: Arc<MessageSpawner>,
     ) -> Box<dyn TabPane> {
